🌱 refactor: ocr 매핑 수정

+++ apps/backend/src/main/java/com/calendarbox/backend/expense/repository/ExpenseOcrTaskRepository.java
@@ -9,7 +9,7 @@ import org.springframework.data.repository.query.Param;
 import java.util.List;
 import java.util.Optional;
 
-public interface ExpenseOcrTaskRepository extends JpaRepository<ExpenseOcrTask, Integer> {
+public interface ExpenseOcrTaskRepository extends JpaRepository<ExpenseOcrTask, Long> {
     Optional<ExpenseOcrTask> findByRequestHash(String requestHash);
 
     @Query(
